+++ src/spnego/__init__.py
@@ -7,23 +7,29 @@ import logging.config
 import os
 
 from spnego._context import ContextReq
+from spnego._credential import (
+    Credential,
+    CredentialCache,
+    KerberosCCache,
+    KerberosKeytab,
+    NTLMHash,
+    Password,
+)
 from spnego.auth import client, server
 from spnego.exceptions import NegotiateOptions
 
 __all__ = [
     'ContextReq',
+    'Credential',
+    'CredentialCache',
+    'KerberosCCache',
+    'KerberosKeytab',
+    'NegotiateOptions',
+    'NTLMHash',
+    'Password',
     'client',
     'server',
-    'NegotiateOptions',
 ]
-
-
-try:
-    from logging import NullHandler
-except ImportError:  # pragma: no cover
-    class NullHandler(logging.Handler):
-        def emit(self, record):
-            pass
 
 
 def _setup_logging(logger):
@@ -37,7 +43,7 @@ def _setup_logging(logger):
         logging.config.dictConfig(config)
     else:
         # no logging was provided
-        logger.addHandler(NullHandler())
+        logger.addHandler(logging.NullHandler())
 
 
 logger = logging.getLogger(__name__)