+++ tests/test_gss.py
@@ -10,7 +10,7 @@ import pytest
 import spnego
 import spnego._gss
 import spnego.iov
-from spnego.exceptions import FeatureMissingError
+from spnego.exceptions import FeatureMissingError, InvalidCredentialError
 
 
 def test_gss_sasl_description_fail(mocker, monkeypatch):
@@ -106,3 +106,18 @@ def test_gssapi_no_kerberos(monkeypatch):
     with pytest.raises(FeatureMissingError, match="The Python gssapi library is not installed so Kerberos cannot be "
                                                   "negotiated."):
         spnego._gss.GSSAPIProxy(None, None, options=spnego.NegotiateOptions.negotiate_kerberos)
+
+
+@pytest.mark.skipif(not spnego._gss.HAS_GSSAPI, reason='Requires the gssapi library to be installed for testing')
+def test_gssapi_no_valid_acceptor_cred():
+    server_creds = [
+        (spnego.KerberosCCache("ccache"), "kerberos"),
+        (spnego.KerberosKeytab("user_princ", "ccache"), "kerberos"),
+        (spnego.NTLMHash("user_princ"), "ntlm"),
+    ]
+    for cred, protocol in server_creds:
+        if protocol not in spnego._gss.GSSAPIProxy.available_protocols():
+            continue
+
+        with pytest.raises(InvalidCredentialError, match="No applicable credentials available"):
+            spnego._gss.GSSAPIProxy(cred, protocol=protocol, usage="accept")